feat: change styling of forms view
